refactor: folders api 모듈화 및 절대경로 적용
