+++ src/api/folders/patchFolder.ts
@@ -1,0 +1,17 @@
+import { ApiModifyFunction, SucceedResponse } from "@/api/const";
+import http from "@/api/http";
+
+export const patchFolder: ApiModifyFunction<SucceedResponse> = (
+  body,
+  onSuccess,
+  onFail,
+  targetId
+) => {
+  return http.patch(
+    `/folders/${targetId}`,
+    { credentials: "include" },
+    body,
+    onSuccess,
+    onFail
+  );
+};
