+++ src/api/folders/postFolder.ts
@@ -1,0 +1,16 @@
+import { ApiModifyFunction, SucceedResponse } from "@/api/const";
+import http from "@/api/http";
+
+export const postFolder: ApiModifyFunction<SucceedResponse> = (
+  body,
+  onSuccess,
+  onFail
+) => {
+  return http.post(
+    `/folders`,
+    { credentials: "include" },
+    body,
+    onSuccess,
+    onFail
+  );
+};
